reorganizes src/app/patterns/ for now
