+++ src/app/patterns/controls/AggregationTypeSelector.svelte
@@ -1,11 +1,11 @@
 <script>
-import FloatingMenu from '../../components/menu/FloatingMenu.svelte';
-import MenuList from '../../components/menu/MenuList.svelte';
-import MenuListItem from '../../components/menu/MenuListItem.svelte';
-import DownCarat from '../../components/icons/DownCarat.svelte';
+import FloatingMenu from '../../../components/menu/FloatingMenu.svelte';
+import MenuList from '../../../components/menu/MenuList.svelte';
+import MenuListItem from '../../../components/menu/MenuListItem.svelte';
+import DownCarat from '../../../components/icons/DownCarat.svelte';
 
-// export let aggregationTypes;
-export let currentView;
+export let aggregationTypes;
+export let currentAggregation;
 export let active;
 
 let button;
@@ -17,22 +17,20 @@ function toggle() {
 }
 
 function setValue(event) {
-  currentView = event.detail.key;
+  currentAggregation = event.detail.key;
   active = false;
 }
 
-export let viewInfo = {
-  explorer: { name: 'Explorer', description: 'Observe trends over time and compare two probes.' },
-  table: { name: 'Table', description: 'Get a table of percentile / proportion values over many builds.' },
+export let aggregationInfo = {
+  avg: { name: 'Average', description: 'Shows the distribution of the average scalar value per client.' },
+  min: { name: 'Minimum', description: 'Shows the distribution of the smallest scalar value per client.' },
+  max: { name: 'Maximum', description: 'Shows the distribution of the largest scalar value per client.' },
+  sum: { name: 'Sum', description: 'Shows the distribution of the sum of scalar values per client.' },
 };
 
 </script>
 
 <style>
-
-div {
-  width: max-content;
-}
 
 .activating-button {
   padding: var(--space-1h);
@@ -68,7 +66,7 @@ div {
 
 <div class=menu-button bind:this={button}>
   <button class=activating-button on:click={toggle} class:active>
-      <div>{viewInfo[currentView].name}</div> <DownCarat size=16 />
+      <div>{aggregationInfo[currentAggregation].name}</div> <DownCarat size=16 />
   </button>
 </div>
 
@@ -81,10 +79,10 @@ div {
     parent={button}
   >
     <MenuList  on:selection={setValue}>
-      {#each Object.entries(viewInfo) as [k, v], i}
-        <MenuListItem  key={k} value={k}>
-          <div class=menu-list-item__title>{v.name}</div>
-          <div class=menu-list-item__description>{v.description}</div>
+      {#each aggregationTypes as agg, i}
+        <MenuListItem  key={agg} value={agg}>
+          <div class=menu-list-item__title>{aggregationInfo[agg].name}</div>
+          <div class=menu-list-item__description>{aggregationInfo[agg].description}</div>
         </MenuListItem>  
       {/each}
     </MenuList>
