+++ src/app/patterns/controls/ProportionMetricTypeControl.svelte
