+++ src/app/patterns/elements/ProportionRow.svelte
@@ -1,7 +1,6 @@
 <script>
 import { createEventDispatcher } from 'svelte';
 import { formatPercent, formatCount } from '../utils/formatters';
-import { percentileLineColorMap } from '../../../../components/data-graphics/utils/color-maps';
 
 import BuildIDCell from '../table/BuildIDCell.svelte';
 import CountProportionCell from '../table/CountProportionCell.svelte';
