+++ src/app/patterns/elements/ProportionTable.svelte
@@ -1,6 +1,6 @@
 <script>
-import DataGraphic from '../../../../components/data-graphics/DataGraphic.svelte';
-import TopAxis from '../../../../components/data-graphics/guides/TopAxis.svelte';
+import DataGraphic from '../../../components/data-graphics/DataGraphic.svelte';
+import TopAxis from '../../../components/data-graphics/guides/TopAxis.svelte';
 import ProportionRow from './ProportionRow.svelte';
 import Pagination from './Pagination.svelte';
