+++ src/app/patterns/elements/ReferenceSymbol.svelte
