+++ src/app/patterns/explorer/ComparisonSummary.svelte
