+++ src/app/patterns/table/BuildIDCell.svelte
