+++ src/app/patterns/table/SingleNumberCell.svelte
