+++ src/app/patterns/utils/animation.js
