+++ src/app/patterns/utils/build-id-utils.js
