+++ src/app/patterns/utils/formatters.js
