+++ stories/data-graphics/GLAM/QuantileTable.svelte
@@ -1,12 +1,10 @@
 <script>
-import QuantileTable from '../../../src/app/patterns/body/elements/QuantileTable.svelte';
+import QuantileTable from '../../../src/app/patterns/elements/QuantileTable.svelte';
 import NAV_URL from '../../../tests/data/browser_engagement_navigation_urlbar_build_id.json';
 import ACTIVE_TICKS from '../../../tests/data/browser_engagement_active_ticks_build_id.json';
 import GCMS from '../../../tests/data/gc_ms_build_id.json';
 
 import { responseToData } from '../../../src/app/state/store';
-
-// import { firefoxVersionMarkers } from '../../../src/app/state/product-versions';
 
 let which = 0;
 // only get the first key
